Вынесет код по работе с точкой маршрута в отдельный презентер
