+++ src/presenter/event-presenter.js
@@ -1,11 +1,10 @@
-import { render, replace } from '../framework/render';
+import { render } from '../framework/render';
 import EventsListView from '../view/events-list-view';
-import WaypointEditView from '../view/waypoint-edit-view';
-import WaypointView from '../view/waypoint-view';
 import SortingView from '../view/sorting-view';
 import EmptyEventsView from '../view/empty-events-view';
 import FailedLoadView from '../view/failed-load-view';
 import LoadingView from '../view/loading-view';
+import WaypointPresenter from './waypoint-presenter';
 
 export default class EventPresenter {
   #containerEl = null;
@@ -44,8 +43,12 @@ export default class EventPresenter {
       return;
     }
 
+    this.#renderWaypointsList();
+  }
+
+  #renderWaypointsList() {
     for (let i = 0; i < this.#waypoints.length; i++) {
-      this.#renderWaypoint(this.#waypoints[i]);
+      this.#renderWaypointPresenter(this.#waypoints[i]);
     }
   }
 
@@ -54,51 +57,12 @@ export default class EventPresenter {
     render(emptyEventsComponent, this.#containerEl);
   }
 
-  #renderWaypoint(waypoint) {
-    const escKeydownHandler = (evt) => {
-      if (evt.key === 'Escape') {
-        evt.preventDefault();
-        replaceEditedWaypointToWaypoint();
-        document.removeEventListener('keydown', escKeydownHandler);
-      }
-    };
-
-    const waypointComponent = new WaypointView({
-      waypoint: waypoint,
+  #renderWaypointPresenter(waypoint) {
+    new WaypointPresenter({
+      waypointsContainerEl: this.#eventsListComponent.element,
       destinations: this.#destinations,
       offers: this.#offers,
-      onBtnUnfoldClick() {
-        replaceWaypointToEditedWaypoint();
-        document.addEventListener('keydown', escKeydownHandler);
-      },
-    });
-
-    const waypointEditComponent = new WaypointEditView({
-      waypoint: waypoint,
-      destinations: this.#destinations,
-      offers: this.#offers,
-      onBtnFoldClick() {
-        replaceEditedWaypointToWaypoint();
-        document.removeEventListener('keydown', escKeydownHandler);
-      },
-      onSubmit() {
-        replaceEditedWaypointToWaypoint();
-        document.removeEventListener('keydown', escKeydownHandler);
-      },
-      onReset() {
-        replaceEditedWaypointToWaypoint();
-        document.removeEventListener('keydown', escKeydownHandler);
-      },
-    });
-
-    function replaceWaypointToEditedWaypoint() {
-      replace(waypointEditComponent, waypointComponent);
-    }
-
-    function replaceEditedWaypointToWaypoint() {
-      replace(waypointComponent, waypointEditComponent);
-    }
-
-    render(waypointComponent, this.#eventsListComponent.element);
+      waypoint,
+    }).init();
   }
 }
